# internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"encoding/json"
	"net/netip"
	"time"

	"github.com/google/uuid"
)

type ApiToken struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Name       string
	TokenHash  []byte
	CreatedAt  time.Time
	ExpiresAt  *time.Time
	RevokedAt  *time.Time
	LastUsedAt *time.Time
}

type AuditLog struct {
	ID         int64
	OccurredAt time.Time
	TenantID   *uuid.UUID
	ActorID    *uuid.UUID
	ActorName  string
	Action     string
	TargetType string
	TargetID   *uuid.UUID
	RequestID  string
	Ip         *netip.Addr
	Details    json.RawMessage
}

type Camera struct {
	ID           uuid.UUID
	TenantID     uuid.UUID
	SiteID       uuid.UUID
	ServerID     uuid.UUID
	RemoteName   string
	DisplayName  string
	Enabled      bool
	Zones        []string
	Lpr          bool
	LiveStream   string
	HqStream     string
	Status       string
	Fps          *float32
	MissingSince *time.Time
	CreatedAt    time.Time
	UpdatedAt    time.Time
	DeletedAt    *time.Time
	DeletedBy    *uuid.UUID
}

type CameraGroup struct {
	ID          uuid.UUID
	TenantID    uuid.UUID
	Name        string
	Description string
	CreatedAt   time.Time
	UpdatedAt   time.Time
	DeletedAt   *time.Time
	DeletedBy   *uuid.UUID
}

type CameraGroupMember struct {
	GroupID  uuid.UUID
	CameraID uuid.UUID
	TenantID uuid.UUID
}

type Event struct {
	ID            uuid.UUID
	TenantID      uuid.UUID
	SiteID        uuid.UUID
	ServerID      uuid.UUID
	CameraID      uuid.UUID
	RemoteID      string
	Severity      string
	Labels        []string
	SubLabels     []string
	Zones         []string
	Audio         []string
	Plates        []string
	DetectionIds  []string
	StartTime     time.Time
	EndTime       *time.Time
	Reviewed      bool
	ThumbKey      string
	ThumbAttempts int32
	FirstSeenAt   time.Time
	UpdatedAt     time.Time
}

type EventSyncState struct {
	ServerID        uuid.UUID
	TenantID        uuid.UUID
	ReviewCursor    *time.Time
	ObjectCursor    *time.Time
	LastSuccessAt   *time.Time
	LastError       string
	BackfilledUntil *time.Time
	UpdatedAt       time.Time
}

type Export struct {
	ID          uuid.UUID
	TenantID    uuid.UUID
	SiteID      uuid.UUID
	ServerID    uuid.UUID
	CameraID    uuid.UUID
	RequestedBy uuid.UUID
	Name        string
	StartTime   time.Time
	EndTime     time.Time
	RemoteID    string
	Status      string
	Progress    float32
	Error       string
	RemotePath  string
	CreatedAt   time.Time
	UpdatedAt   time.Time
	DeletedAt   *time.Time
}

type FrigateServer struct {
	ID             uuid.UUID
	TenantID       uuid.UUID
	SiteID         uuid.UUID
	Name           string
	BaseUrl        string
	Username       string
	PasswordSealed []byte
	TlsSkipVerify  bool
	FrigateVersion string
	Capabilities   json.RawMessage
	Status         string
	LastSeenAt     *time.Time
	LastCheckedAt  *time.Time
	LastError      string
	Stats          json.RawMessage
	CreatedAt      time.Time
	UpdatedAt      time.Time
	DeletedAt      *time.Time
	DeletedBy      *uuid.UUID
	AuthMode       string
}

type LprRead struct {
	ID              uuid.UUID
	TenantID        uuid.UUID
	SiteID          uuid.UUID
	ServerID        uuid.UUID
	CameraID        uuid.UUID
	RemoteEventID   string
	Plate           string
	PlateNormalized string
	Score           *float32
	Label           string
	SubLabel        string
	Zones           []string
	SeenAt          time.Time
	FirstSeenAt     time.Time
	UpdatedAt       time.Time
}

type PermissionGrant struct {
	ID          uuid.UUID
	TenantID    *uuid.UUID
	SubjectType string
	SubjectID   uuid.UUID
	Permission  string
	Effect      string
	ScopeType   string
	ScopeID     *uuid.UUID
	CreatedBy   *uuid.UUID
	CreatedAt   time.Time
}

type Session struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	TokenHash  []byte
	CreatedAt  time.Time
	LastSeenAt time.Time
	ExpiresAt  time.Time
	RevokedAt  *time.Time
	Ip         *netip.Addr
	UserAgent  string
}

type Site struct {
	ID        uuid.UUID
	TenantID  uuid.UUID
	Name      string
	Timezone  string
	Address   string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt *time.Time
	DeletedBy *uuid.UUID
}

type Tenant struct {
	ID        uuid.UUID
	Slug      string
	Name      string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt *time.Time
	DeletedBy *uuid.UUID
}

type User struct {
	ID                 uuid.UUID
	TenantID           *uuid.UUID
	Username           string
	Email              *string
	DisplayName        string
	Status             string
	CreatedAt          time.Time
	UpdatedAt          time.Time
	DeletedAt          *time.Time
	DeletedBy          *uuid.UUID
	PasswordHash       string
	PasswordChangedAt  *time.Time
	MustChangePassword bool
	MfaSecretSealed    []byte
	MfaEnabled         bool
	FailedLogins       int32
	LockedUntil        *time.Time
	LastLoginAt        *time.Time
}

type UserGroup struct {
	ID          uuid.UUID
	TenantID    *uuid.UUID
	Name        string
	Description string
	CreatedAt   time.Time
	UpdatedAt   time.Time
	DeletedAt   *time.Time
	DeletedBy   *uuid.UUID
}

type UserGroupMember struct {
	GroupID uuid.UUID
	UserID  uuid.UUID
}

type View struct {
	ID        uuid.UUID
	TenantID  uuid.UUID
	OwnerID   uuid.UUID
	Name      string
	Shared    bool
	Layout    json.RawMessage
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt *time.Time
}