# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: OpenVMS API
  version: 0.3.0
  description: |
    Control plane API for the Federated Frigate VMS.
    Every endpoint is defined here first; server and client code are generated from this file.
servers:
  - url: /
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/me/password:
    post:
      tags: [identity]
      operationId: changePassword
      summary: Change your own password
      description: Other sessions of the user are closed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 10 }
      responses:
        "204": { description: Changed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/mfa/setup:
    post:
      tags: [identity]
      operationId: setupMfa
      summary: Start TOTP enrolment
      description: Returns a new secret. MFA is only enforced after enableMfa confirms a code.
      responses:
        "200":
          description: Secret to add to an authenticator app
          content:
            application/json:
              schema:
                type: object
                required: [secret, otpauth_url]
                properties:
                  secret: { type: string, description: Base32 secret for manual entry. }
                  otpauth_url: { type: string, description: otpauth:// URL (render as QR). }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/mfa/enable:
    post:
      tags: [identity]
      operationId: enableMfa
      summary: Confirm a TOTP code and turn MFA on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, pattern: "^[0-9 ]{6,7}$" }
      responses:
        "204": { description: Enabled }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/mfa/disable:
    post:
      tags: [identity]
      operationId: disableMfa
      summary: Turn MFA off (needs the password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password }
      responses:
        "204": { description: Disabled }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/users:
    get:
      tags: [identity]
      operationId: listUsers
      summary: List users (users.view)
      parameters:
        - { name: tenant_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [identity]
      operationId: createUser
      summary: Create a user (users.manage; groups.manage to add groups)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/users/{userId}:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [identity]
      operationId: getUser
      summary: Get a user
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [identity]
      operationId: updateUser
      summary: Update, disable, reset the password or MFA of a user (users.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [identity]
      operationId: deleteUser
      summary: Remove a user (soft delete, sessions closed)
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/groups:
    get:
      tags: [identity]
      operationId: listUserGroups
      summary: List user groups (groups.view)
      parameters:
        - { name: tenant_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [identity]
      operationId: createUserGroup
      summary: Create a user group (groups.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserGroupInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/groups/{groupId}:
    parameters:
      - { name: groupId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [identity]
      operationId: getUserGroup
      summary: Get a user group
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [identity]
      operationId: replaceUserGroup
      summary: Rename a group and replace its members
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserGroupInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserGroup" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [identity]
      operationId: deleteUserGroup
      summary: Delete a user group (its grants stop applying)
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/audit:
    get:
      tags: [audit]
      operationId: listAudit
      summary: Read the audit trail (audit.view)
      parameters:
        - { name: tenant_id, in: query, schema: { type: string, format: uuid } }
        - { name: action, in: query, schema: { type: string } }
        - { name: actor_id, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: before_id, in: query, schema: { type: integer, format: int64 }, description: Return entries older than this id (pagination). }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Entries, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for scripts and integrations. Browsers use the openvms_session cookie set by login.
    cookieAuth:
      type: apiKey
      in: cookie
      name: openvms_session
      description: Browser session. State-changing requests must also send the X-OpenVMS-Request header.
  responses:
    BadRequest:
      description: The request is invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The caller lacks the permission on this resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The resource does not exist or belongs to another tenant
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The request conflicts with existing data
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    UserStatus:
      type: string
      enum: [active, disabled, locked, pending]
    User:
      type: object
      required: [id, tenant_id, username, display_name, status, mfa_enabled, has_password, must_change_password, group_ids, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid, nullable: true, description: Null for platform users. }
        username: { type: string }
        display_name: { type: string }
        email: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/UserStatus" }
        mfa_enabled: { type: boolean }
        has_password: { type: boolean, description: False for users that only use API tokens. }
        must_change_password: { type: boolean }
        locked_until: { type: string, format: date-time, nullable: true }
        last_login_at: { type: string, format: date-time, nullable: true }
        group_ids: { type: array, items: { type: string, format: uuid } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserCreate:
      type: object
      required: [username]
      properties:
        tenant_id:
          type: string
          format: uuid
          description: Defaults to the caller's tenant. Omit (platform administrators) for a platform user.
        username: { type: string, minLength: 3, maxLength: 64, pattern: "^[A-Za-z0-9._@-]+$" }
        display_name: { type: string, maxLength: 200 }
        email: { type: string, maxLength: 200 }
        password: { type: string, format: password, minLength: 10, description: Omit for a user that only uses API tokens. }
        must_change_password: { type: boolean, default: false }
        group_ids: { type: array, items: { type: string, format: uuid } }
    UserUpdate:
      type: object
      properties:
        display_name: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, maxLength: 200 }
        status: { type: string, enum: [active, disabled, locked] }
        password: { type: string, format: password, minLength: 10, description: Resets the password and closes the user's sessions. }
        must_change_password: { type: boolean }
        group_ids: { type: array, items: { type: string, format: uuid } }
        disable_mfa: { type: boolean, description: Removes the second factor (lost phone). }
    UserGroupInput:
      type: object
      required: [name]
      properties:
        tenant_id: { type: string, format: uuid, description: Defaults to the caller's tenant. }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 1000 }
        member_ids: { type: array, items: { type: string, format: uuid }, maxItems: 5000 }
    UserGroup:
      type: object
      required: [id, tenant_id, name, description, member_ids, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        description: { type: string }
        member_ids: { type: array, items: { type: string, format: uuid } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AuditEntry:
      type: object
      required: [id, occurred_at, actor_name, action, target_type, request_id, details]
      properties:
        id: { type: integer, format: int64 }
        occurred_at: { type: string, format: date-time }
        tenant_id: { type: string, format: uuid, nullable: true }
        actor_id: { type: string, format: uuid, nullable: true }
        actor_name: { type: string }
        action: { type: string, example: LOGIN_SUCCESS }
        target_type: { type: string }
        target_id: { type: string, format: uuid, nullable: true }
        request_id: { type: string }
        ip: { type: string, nullable: true }
        details: { type: object, additionalProperties: true }
...